Overall refactoring and bug fixes - Separate the metrics functions from predictors.rs to spatial_metrics.rs - [features.rs] Fix a bug on the eigenvectors_b writing on local_features - [features.rs] Test the try_svd function with the lowest possible epsilon (experimental) - Overall polishing
